Change the input of Motors from the scale of 0->1 to 0->255
Change the iput type to int from float and do not scale the command
within the Motror driver.

+++ arduino/include/Motor.h
@@ -8,7 +8,7 @@ class Motor
 public:
     Motor(int fwd_pin, int bck_pin, int max_output = 255, int min_output = 0);
 
-    int16_t setSpeed(double vel);
+    int16_t setSpeed(int16_t vel);
 
 private:
     const int fwd_pin_;
